Move tryWithResourceCleanUp into jdt.core.manipulation
